refactor: split compiler modules and refresh checks
