+++ crates/vize/src/commands/check/runner.rs
@@ -12,12 +12,11 @@ use std::{
     time::{Duration, Instant},
 };
 
-use ignore::WalkBuilder;
 use vize_canon::{
     batch::TypeChecker as BatchTypeCheckerTrait, BatchTypeChecker, BatchTypeCheckerOptions,
     DeclarationEmitOptions,
 };
-use vize_carton::{cstr, profiler::global_profiler, FxHashSet, String};
+use vize_carton::{cstr, profiler::global_profiler, String};
 
 use crate::commands::profile::{
     print_profile_report, ProfilePhase, ProfilePhaseKind, ProfileReport,
@@ -29,221 +28,13 @@ use super::{
     CheckArgs,
 };
 
-/// Run type checking via Unix socket connection to check-server.
+mod collect;
 #[cfg(unix)]
-pub(crate) fn run_with_socket(args: &CheckArgs, socket_path: &str) {
-    use std::{
-        io::{BufRead, BufReader, Write},
-        os::unix::net::UnixStream,
-    };
+mod socket;
 
-    use super::{JsonRpcResponse, ServerCheckResult};
-
-    let start = Instant::now();
-    let collect_start = Instant::now();
-    #[allow(clippy::disallowed_types)]
-    let default_patterns = vec![std::string::String::from(".")];
-    let files = if args.patterns.is_empty() {
-        collect_vue_files(&default_patterns)
-    } else {
-        collect_vue_files(&args.patterns)
-    };
-    let collect_time = collect_start.elapsed();
-
-    if files.is_empty() {
-        eprintln!("No .vue files found matching inputs: {:?}", args.patterns);
-        return;
-    }
-
-    let connect_start = Instant::now();
-    let mut stream = match UnixStream::connect(socket_path) {
-        Ok(stream) => stream,
-        Err(error) => {
-            eprintln!(
-                "\x1b[31mError:\x1b[0m Failed to connect to check-server: {}",
-                error
-            );
-            eprintln!();
-            eprintln!("\x1b[33mHint:\x1b[0m Start the server first:");
-            eprintln!("  vize check-server --socket {}", socket_path);
-            std::process::exit(1);
-        }
-    };
-    let connect_time = connect_start.elapsed();
-
-    if !args.quiet {
-        eprintln!("Connected to check-server at {}", socket_path);
-        eprintln!("Type checking {} Vue files...", files.len());
-    }
-
-    let mut total_errors = 0usize;
-    #[allow(clippy::disallowed_types, clippy::disallowed_methods)]
-    let mut results: Vec<(std::string::String, ServerCheckResult)> = Vec::new();
-
-    let request_start = Instant::now();
-    for path in &files {
-        #[allow(clippy::disallowed_types)]
-        let source = match fs::read_to_string(path) {
-            Ok(source) => source,
-            Err(error) => {
-                eprintln!("Failed to read {}: {}", path.display(), error);
-                continue;
-            }
-        };
-
-        #[allow(clippy::disallowed_methods)]
-        let filename = path.to_string_lossy().to_string();
-
-        let request = serde_json::json!({
-            "jsonrpc": "2.0",
-            "id": 1,
-            "method": "check",
-            "params": {
-                "uri": filename,
-                "content": source,
-            }
-        });
-
-        if writeln!(stream, "{}", request).is_err() || stream.flush().is_err() {
-            eprintln!("Failed to send request");
-            break;
-        }
-
-        let mut reader = BufReader::new(&stream);
-        #[allow(clippy::disallowed_types)]
-        let mut response_line = std::string::String::new();
-        if reader.read_line(&mut response_line).is_err() {
-            eprintln!("Failed to read response");
-            break;
-        }
-
-        let response: JsonRpcResponse = match serde_json::from_str(&response_line) {
-            Ok(response) => response,
-            Err(error) => {
-                eprintln!("Failed to parse response: {}", error);
-                continue;
-            }
-        };
-
-        if let Some(error) = response.error {
-            eprintln!("Server error: {}", error.message);
-            continue;
-        }
-
-        if let Some(result) = response.result {
-            total_errors += result.error_count;
-            if args.show_virtual_ts {
-                eprintln!("\n=== {} ===", filename);
-                eprintln!("{}", result.virtual_ts);
-            }
-            results.push((filename, result));
-        }
-    }
-    let request_time = request_start.elapsed();
-
-    let render_start = Instant::now();
-    if !args.quiet {
-        for (filename, result) in &results {
-            if result.diagnostics.is_empty() {
-                continue;
-            }
-            println!("\n\x1b[4m{}\x1b[0m", filename);
-            for diagnostic in &result.diagnostics {
-                let color = if diagnostic.severity == "error" {
-                    "\x1b[31m"
-                } else {
-                    "\x1b[33m"
-                };
-                let code = diagnostic
-                    .code
-                    .as_ref()
-                    .map(|code| cstr!(" [{}]", code))
-                    .unwrap_or_default();
-                println!(
-                    "  {}{}:{}:{}\x1b[0m{} {}",
-                    color,
-                    diagnostic.severity,
-                    diagnostic.line,
-                    diagnostic.column,
-                    code,
-                    diagnostic.message
-                );
-            }
-        }
-    }
-    let render_time = render_start.elapsed();
-    let total_time = start.elapsed();
-
-    let status = if total_errors > 0 {
-        "\x1b[31m\u{2717}\x1b[0m"
-    } else {
-        "\x1b[32m\u{2713}\x1b[0m"
-    };
-    println!(
-        "\n{} Type checked {} Vue files in {:.2?} (via socket)",
-        status,
-        files.len(),
-        total_time
-    );
-    if args.profile {
-        let phases = [
-            ProfilePhase {
-                name: "collect files",
-                duration: collect_time,
-                kind: ProfilePhaseKind::Wall,
-                note: "Vue input discovery",
-            },
-            ProfilePhase {
-                name: "connect socket",
-                duration: connect_time,
-                kind: ProfilePhaseKind::Wall,
-                note: "Unix socket handshake",
-            },
-            ProfilePhase {
-                name: "request checks",
-                duration: request_time,
-                kind: ProfilePhaseKind::Wall,
-                note: "read, send, receive",
-            },
-            ProfilePhase {
-                name: "render diagnostics",
-                duration: render_time,
-                kind: ProfilePhaseKind::Wall,
-                note: "terminal output",
-            },
-        ];
-        let mut recommendations: Vec<String> = Vec::new();
-        if request_time > connect_time * 4 {
-            recommendations.push(
-                "Socket request time dominates; profile the running check-server process next."
-                    .into(),
-            );
-        }
-        let summary = cstr!(
-            "{} Vue file(s), {} error(s), socket {}",
-            files.len(),
-            total_errors,
-            socket_path
-        );
-        let report = ProfileReport {
-            title: "check --socket",
-            summary: summary.as_str(),
-            total: total_time,
-            phases: &phases,
-            files: &[],
-            slow_threshold: Duration::from_millis(0),
-            throughput_bytes: None,
-            operations: None,
-            recommendations: &recommendations,
-        };
-        print_profile_report(&report);
-    }
-    if total_errors > 0 {
-        println!("  \x1b[31m{} error(s)\x1b[0m", total_errors);
-        std::process::exit(1);
-    }
-    println!("  \x1b[32mNo type errors found!\x1b[0m");
-}
+use collect::collect_check_files;
+#[cfg(unix)]
+pub(crate) use socket::run_with_socket;
 
 /// Run type checking directly with a materialized Corsa project.
 pub(crate) fn run_direct(args: &CheckArgs) {
@@ -768,176 +559,10 @@ fn find_nearest_tsconfig_dir(path: &Path) -> Option<PathBuf> {
     None
 }
 
-#[allow(clippy::disallowed_types)]
-pub(crate) fn collect_check_files(patterns: &[std::string::String]) -> Vec<PathBuf> {
-    let mut files = Vec::new();
-    let mut seen = FxHashSet::default();
-
-    for pattern in patterns {
-        let candidate = PathBuf::from(pattern);
-        if candidate.exists() {
-            if candidate.is_file() {
-                let candidate = normalize_input_path(&candidate);
-                if is_supported_check_file(&candidate) && seen.insert(candidate.clone()) {
-                    files.push(candidate);
-                }
-                continue;
-            }
-            if candidate.is_dir() {
-                collect_from_dir(&candidate, &mut files, &mut seen);
-                continue;
-            }
-        }
-
-        let base_dir = base_dir_from_pattern(pattern);
-        collect_from_dir(base_dir.as_path(), &mut files, &mut seen);
-    }
-
-    files.sort();
-    files
-}
-
-#[allow(clippy::disallowed_types)]
-pub(crate) fn collect_vue_files(patterns: &[std::string::String]) -> Vec<PathBuf> {
-    let mut files = Vec::new();
-    let mut seen = FxHashSet::default();
-
-    for pattern in patterns {
-        let candidate = PathBuf::from(pattern);
-        if candidate.exists() {
-            if candidate.is_file() {
-                let candidate = normalize_input_path(&candidate);
-                if candidate
-                    .extension()
-                    .and_then(|extension| extension.to_str())
-                    == Some("vue")
-                    && seen.insert(candidate.clone())
-                {
-                    files.push(candidate);
-                }
-                continue;
-            }
-            if candidate.is_dir() {
-                collect_from_dir_filtered(&candidate, &mut files, &mut seen, true);
-                continue;
-            }
-        }
-
-        let base_dir = base_dir_from_pattern(pattern);
-        collect_from_dir_filtered(&base_dir, &mut files, &mut seen, true);
-    }
-
-    files.sort();
-    files
-}
-
-fn collect_from_dir(dir: &Path, files: &mut Vec<PathBuf>, seen: &mut FxHashSet<PathBuf>) {
-    collect_from_dir_filtered(dir, files, seen, false);
-}
-
-fn collect_from_dir_filtered(
-    dir: &Path,
-    files: &mut Vec<PathBuf>,
-    seen: &mut FxHashSet<PathBuf>,
-    vue_only: bool,
-) {
-    let skip_generated = should_skip_generated_for_root(dir);
-    let walker = WalkBuilder::new(dir)
-        .standard_filters(true)
-        .hidden(true)
-        .build_parallel();
-
-    let collected = std::sync::Mutex::new(Vec::<PathBuf>::new());
-    walker.run(|| {
-        let collected = &collected;
-        Box::new(move |entry| {
-            if let Ok(entry) = entry {
-                let path = entry.path();
-                if path.is_file()
-                    && is_supported_collect_file(path, vue_only)
-                    && (!skip_generated || !is_generated_path(path))
-                {
-                    if let Ok(mut collected) = collected.lock() {
-                        collected.push(path.to_path_buf());
-                    }
-                }
-            }
-            ignore::WalkState::Continue
-        })
-    });
-
-    let Ok(collected) = collected.into_inner() else {
-        return;
-    };
-    for path in collected {
-        let path = normalize_input_path(&path);
-        if seen.insert(path.clone()) {
-            files.push(path);
-        }
-    }
-}
-
-fn base_dir_from_pattern(pattern: &str) -> PathBuf {
-    let glob_start = pattern.find(['*', '?', '[', '{']).unwrap_or(pattern.len());
-    let prefix = &pattern[..glob_start];
-    let base = if prefix.is_empty() {
-        "."
-    } else if let Some(index) = prefix.rfind('/') {
-        &prefix[..index]
-    } else {
-        prefix
-    };
-    if base.is_empty() {
-        PathBuf::from(".")
-    } else {
-        PathBuf::from(base)
-    }
-}
-
-fn normalize_input_path(path: &Path) -> PathBuf {
-    path.canonicalize().unwrap_or_else(|_| path.to_path_buf())
-}
-
-fn should_skip_generated_for_root(root: &Path) -> bool {
-    !root
-        .components()
-        .any(|component| component.as_os_str().to_str() == Some("__agent_only"))
-}
-
 fn display_path(base: &Path, path: &Path) -> vize_carton::String {
     path.strip_prefix(base)
         .map(|relative| cstr!("{}", relative.display()))
         .unwrap_or_else(|_| cstr!("{}", path.display()))
-}
-
-fn is_generated_path(path: &Path) -> bool {
-    path.components().any(|component| {
-        component
-            .as_os_str()
-            .to_str()
-            .is_some_and(|name| matches!(name, "__agent_only" | "target"))
-    })
-}
-
-fn is_supported_collect_file(path: &Path, vue_only: bool) -> bool {
-    if vue_only {
-        return path.extension().and_then(|extension| extension.to_str()) == Some("vue");
-    }
-    is_supported_check_file(path)
-}
-
-fn is_supported_check_file(path: &Path) -> bool {
-    if path
-        .file_name()
-        .and_then(|name| name.to_str())
-        .is_some_and(|name| name.ends_with(".d.ts"))
-    {
-        return true;
-    }
-
-    path.extension()
-        .and_then(|extension| extension.to_str())
-        .is_some_and(|extension| matches!(extension, "vue" | "ts" | "tsx" | "mts" | "cts"))
 }
 
 /// Parse a `.d.ts` file containing `ComponentCustomProperties` augmentation.
@@ -961,66 +586,8 @@ fn parse_dts_globals(
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        base_dir_from_pattern, collect_check_files, collect_vue_files, resolve_declaration_dir,
-    };
-    use std::fs;
+    use super::resolve_declaration_dir;
     use std::path::{Path, PathBuf};
-    use vize_carton::cstr;
-
-    fn unique_case_dir(name: &str) -> PathBuf {
-        static NEXT_CASE_ID: std::sync::atomic::AtomicUsize =
-            std::sync::atomic::AtomicUsize::new(0);
-        let case_id = NEXT_CASE_ID.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-        Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("__agent_only")
-            .join("tests")
-            .join(cstr!("{name}-{}-{case_id}", std::process::id()).as_str())
-    }
-
-    #[test]
-    fn base_dir_from_glob_patterns() {
-        assert_eq!(
-            base_dir_from_pattern("./src/**/*.vue"),
-            PathBuf::from("./src")
-        );
-        assert_eq!(base_dir_from_pattern("."), PathBuf::from("."));
-    }
-
-    #[test]
-    fn collect_check_files_includes_ts_and_vue_and_dts() {
-        let case_dir = unique_case_dir("collect-check");
-        let _ = fs::remove_dir_all(&case_dir);
-        fs::create_dir_all(case_dir.join("src")).unwrap();
-        fs::write(case_dir.join("src/App.vue"), "").unwrap();
-        fs::write(case_dir.join("src/main.ts"), "").unwrap();
-        fs::write(case_dir.join("src/env.d.ts"), "").unwrap();
-        fs::write(case_dir.join("src/skip.js"), "").unwrap();
-
-        let files = collect_check_files(&vec![case_dir.display().to_string()]);
-
-        assert_eq!(files.len(), 3);
-        assert!(files.iter().any(|path| path.ends_with("App.vue")));
-        assert!(files.iter().any(|path| path.ends_with("main.ts")));
-        assert!(files.iter().any(|path| path.ends_with("env.d.ts")));
-
-        let _ = fs::remove_dir_all(&case_dir);
-    }
-
-    #[test]
-    fn collect_vue_files_stays_vue_only() {
-        let case_dir = unique_case_dir("collect-vue");
-        let _ = fs::remove_dir_all(&case_dir);
-        fs::create_dir_all(case_dir.join("src")).unwrap();
-        fs::write(case_dir.join("src/App.vue"), "").unwrap();
-        fs::write(case_dir.join("src/main.ts"), "").unwrap();
-
-        let files = collect_vue_files(&vec![case_dir.display().to_string()]);
-
-        assert_eq!(files, vec![case_dir.join("src/App.vue")]);
-
-        let _ = fs::remove_dir_all(&case_dir);
-    }
 
     #[test]
     fn resolve_declaration_dir_defaults_to_dist_types() {
